Добавил более подробное логирование в объектах
Оптимизировал перехват ошибок в классе public

+++ src/objects/public.py
@@ -1,10 +1,9 @@
-import dotenv
-
 import os
 
 from src.objects.interceptor import Interceptor
 from src.objects.video_queue import VideoQueue, DebugVideoQueue
 from src.settings import Settings
+from src.logger import pub_logger
 from src.my_exceptions import (OverOneStartedException, QueueLenException,
                                AccessDeniedException, NoFoundVideoException,
                                NoValidInterPublicException, NoValidVideoPathException,
@@ -28,82 +27,73 @@ class Public:
         self.stop = False
     
     def synchronize(self):
-        list_dir = os.listdir(f'.{self.SL}{self.settings.VIDEO_PATH[0:-1]}')
-        
-        if not self.inter_public in list_dir:
-            try:
+        try:
+            list_dir = os.listdir(f'.{self.SL}{self.settings.VIDEO_PATH[0:-1]}')
+            
+            if not self.inter_public in list_dir:
                 self.interceptor.intercept_video()
-            except AccessDeniedException:
-                raise AccessDeniedException(public_id=self.public_id)
+                
+            list_media = os.listdir(f'.{self.SL}{self.settings.VIDEO_PATH}{self.inter_public}/')
+            diff = self.settings.MAX_LEN_QUEUE - len(list_media)
             
-            except NoValidInterPublicException:
-                raise NoValidInterPublicException(public_id=self.public_id)
-            
-        list_media = os.listdir(f'.{self.SL}{self.settings.VIDEO_PATH}{self.inter_public}/')
-        diff = self.settings.MAX_LEN_QUEUE - len(list_media)
-        
-        for i in range(diff):
-            try:
+            for i in range(diff):
                 self.interceptor.intercept_video()
-            except AccessDeniedException:
-                raise AccessDeniedException(public_id=self.public_id)
-            
-            except NoValidInterPublicException:
-                raise NoValidInterPublicException(public_id=self.public_id)
-                 
-        for i in os.listdir(f'.{self.SL}{self.settings.VIDEO_PATH}{self.inter_public}/'):
-            i = i.replace('.mp4', '')
-            
-            if not i in self.video_queue.queue:       
-                self.video_queue.add_video(i)         
                         
+            for i in os.listdir(f'.{self.SL}{self.settings.VIDEO_PATH}{self.inter_public}/'):
+                i = i.replace('.mp4', '')
+                
+                if not i in self.video_queue.queue:       
+                    self.video_queue.add_video(i)
+                             
+        except AccessDeniedException:
+                    raise AccessDeniedException(public_id=self.public_id)
+        except NoValidInterPublicException:
+                    raise NoValidInterPublicException(public_id=self.public_id)
+                              
     async def start(self):
-        if not self.started:
-            self.synchronize()
-            self.started = True
-            try:
+        try:
+            if not self.started:
+                pub_logger.info(f'Паблик {self.public_id} запущен.')
+                
+                self.synchronize()
+                self.started = True
+                
                 video = await self.video_queue.run_next_video(self.inter_public, 
-                                                          self.public_id)
-            except NoValidOwnPublicException:
-                self.started = False
-                raise NoValidOwnPublicException(public_id=self.public_id)
-            except NoValidVideoPathException:
-                self.started = False
-                raise NoValidVideoPathException(public_id=self.public_id)
-            if video:
-                os.remove(f'.{self.SL}{self.settings.VIDEO_PATH}{self.inter_public}{self.SL}{video}.mp4')
-            
-            while True:
-                if not self.stop:
-                    try:
+                                                                self.public_id)
+                if video:
+                    os.remove(f'.{self.SL}{self.settings.VIDEO_PATH}{self.inter_public}{self.SL}{video}.mp4')
+                
+                while True:
+                    if not self.stop:
+                        pub_logger.info(f'Работа цикла у паблика {self.public_id}')
+                        
                         self.video_queue.add_video(str(self.interceptor.intercept_video()))
                         
-                    except NoValidInterPublicException:
-                        self.started = False
-                        raise NoValidInterPublicException(public_id=self.public_id)
-                    except AccessDeniedException:
-                        self.started = False
-                        raise AccessDeniedException(public_id=self.public_id)
-                    
-                    try:
                         video = await self.video_queue.run_next_video(self.inter_public, 
-                                                                      self.public_id)
-                    except NoValidOwnPublicException:
+                                                                        self.public_id)
+                        
+                        if video:
+                            os.remove(f'.{self.SL}{self.settings.VIDEO_PATH}{self.inter_public}{self.SL}{video}.mp4')
+                    else:
                         self.started = False
-                        raise NoValidOwnPublicException(public_id=self.public_id) 
-                    except NoValidVideoPathException:
-                        self.started = False
-                        raise NoValidVideoPathException(public_id=self.public_id)
-                    
-                    if video:
-                        os.remove(f'.{self.SL}{self.settings.VIDEO_PATH}{self.inter_public}{self.SL}{video}.mp4')
-                else:
-                    self.started = False
-                    self.stop = False
-                    break
-        else:
-            raise OverOneStartedException
-    
+                        self.stop = False
+                        break
+            else:
+                raise OverOneStartedException
+            
+        except NoValidOwnPublicException:
+            self.started = False
+            raise NoValidOwnPublicException(public_id=self.public_id)
+        except NoValidVideoPathException:
+            self.started = False
+            raise NoValidVideoPathException(public_id=self.public_id)
+        except NoValidInterPublicException:
+            self.started = False
+            raise NoValidInterPublicException(public_id=self.public_id)
+        except AccessDeniedException:
+            self.started = False
+            raise AccessDeniedException(public_id=self.public_id)
+                      
     def add_video(self, video_id: str):
         if len(self.video_queue.queue) < self.settings.MAX_LEN_QUEUE:
             
@@ -138,47 +128,44 @@ class DebugPublic(Public):
             self.stop = False
         
         def start(self):
-            if not self.started:
-                self.synchronize()
-                
-                self.started = True
-                try:
+            try:
+                if not self.started:
+                    pub_logger.info(f'Паблик {self.public_id} запущен.')
+                    
+                    self.synchronize()
+                    self.started = True
+                    
                     video = self.video_queue.run_next_video(self.inter_public, 
                                                             self.public_id)
-                except NoValidOwnPublicException:
-                    self.started = False
-                    raise NoValidOwnPublicException(public_id=self.public_id)
-                except NoValidVideoPathException:
-                    self.started = False
-                    raise NoValidVideoPathException(public_id=self.public_id)
-                
-                os.remove(f'.{self.SL}{self.settings.VIDEO_PATH}{self.inter_public}{self.SL}{video}.mp4')
-                
-                while True:
-                    if not self.stop:
-                        try:
+                    if video:   
+                        os.remove(f'.{self.SL}{self.settings.VIDEO_PATH}{self.inter_public}{self.SL}{video}.mp4')
+                        
+                    while True:
+                        if not self.stop:
+                            pub_logger.info(f'Работа цикла у паблика {self.public_id}')
+                            
                             self.video_queue.add_video(str(self.interceptor.intercept_video()))
                             
-                        except NoValidInterPublicException:
-                            self.started = False
-                            raise NoValidInterPublicException(public_id=self.public_id)
-                        except AccessDeniedException:
-                            self.started = False
-                            raise AccessDeniedException(public_id=self.public_id)
-                        
-                        try:
                             video = self.video_queue.run_next_video(self.inter_public, 
                                                                     self.public_id)
-                        except NoValidOwnPublicException:
+                            if video:
+                                os.remove(f'.{self.SL}{self.settings.VIDEO_PATH}{self.inter_public}{self.SL}{video}.mp4')
+                        else:
                             self.started = False
-                            raise NoValidOwnPublicException(public_id=self.public_id) 
-                        except NoValidVideoPathException:
-                            self.started = False
-                            raise NoValidVideoPathException(public_id=self.public_id)
-                        
-                        os.remove(f'.{self.SL}{self.settings.VIDEO_PATH}{self.inter_public}{self.SL}{video}.mp4')
-                    else:
-                        self.started = False
-                        break
-            else:
-                raise OverOneStartedException
+                            self.stop = False
+                            break
+                else:
+                    raise OverOneStartedException
+                
+            except NoValidOwnPublicException:
+                self.started = False
+                raise NoValidOwnPublicException(public_id=self.public_id)
+            except NoValidVideoPathException:
+                self.started = False
+                raise NoValidVideoPathException(public_id=self.public_id)
+            except NoValidInterPublicException:
+                self.started = False
+                raise NoValidInterPublicException(public_id=self.public_id)
+            except AccessDeniedException:
+                self.started = False
+                raise AccessDeniedException(public_id=self.public_id)
